feat(finance): introduce financial service layer
