+++ app/build.gradle
@@ -29,7 +29,7 @@ android {
         minSdk 26
         targetSdk 35
         versionCode 23
-        versionName "0.23.12"
+        versionName "0.23.13"
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
         buildConfigField "String", "FIREBASE_API_KEY", "\"${firebaseKey}\""
     }
