+++ app/src/androidTest/java/gr/eduinvoice/ui/invoice/InvoiceScreenTest.kt
@@ -81,6 +81,7 @@ class InvoiceScreenTest {
             override fun getLessonsWithStudentsByStudent(studentId: Long, userId: Long): Flow<List<LessonWithStudent>> = lessonFlow.asStateFlow()
             override fun getLessonsWithStudentsInDateRange(startDate: String, endDate: String, userId: Long): Flow<List<LessonWithStudent>> = lessonFlow.asStateFlow()
             override fun getLessonsWithStudentsByStudentAndDateRange(studentId: Long, startDate: String, endDate: String, userId: Long): Flow<List<LessonWithStudent>> = lessonFlow.asStateFlow()
+            override suspend fun insertGroupLessons(lessons: List<Lesson>): List<Long> = lessons.map { it.id }
         }
         val groupDao = object : GroupDao {
             override suspend fun insertGroup(group: StudentGroup): Long = 1L
