+++ app/src/androidTest/java/gr/eduinvoice/ui/student/StudentScreenTest.kt
@@ -76,6 +76,7 @@ class StudentScreenTest {
             override fun getLessonsWithStudentsByStudent(studentId: Long, userId: Long): Flow<List<LessonWithStudent>> = flowOf(emptyList<LessonWithStudent>())
             override fun getLessonsWithStudentsInDateRange(startDate: String, endDate: String, userId: Long): Flow<List<LessonWithStudent>> = flowOf(emptyList<LessonWithStudent>())
             override fun getLessonsWithStudentsByStudentAndDateRange(studentId: Long, startDate: String, endDate: String, userId: Long): Flow<List<LessonWithStudent>> = flowOf(emptyList<LessonWithStudent>())
+            override suspend fun insertGroupLessons(lessons: List<Lesson>): List<Long> = lessons.map { it.id }
         }
         val groupDao = object : GroupDao {
             override suspend fun insertGroup(group: StudentGroup): Long = 1L
