+++ app/src/test/java/gr/eduinvoice/ui/lesson/LessonViewModelGroupTest.kt
@@ -207,6 +207,11 @@ class LessonViewModelGroupTest {
         override fun getLessonsWithStudentsByStudent(studentId: Long, userId: Long): Flow<List<gr.eduinvoice.data.database.LessonWithStudent>> = flowOf(emptyList())
         override fun getLessonsWithStudentsInDateRange(startDate: String, endDate: String, userId: Long): Flow<List<gr.eduinvoice.data.database.LessonWithStudent>> = flowOf(emptyList())
         override fun getLessonsWithStudentsByStudentAndDateRange(studentId: Long, startDate: String, endDate: String, userId: Long): Flow<List<gr.eduinvoice.data.database.LessonWithStudent>> = flowOf(emptyList())
+
+        override suspend fun insertGroupLessons(lessons: List<Lesson>): List<Long> {
+            lessons.forEach { insert(it) }
+            return lessons.map { it.id }
+        }
     }
 
     class FakeGroupDao(
