+++ app/src/test/java/gr/eduinvoice/ui/lessons/LessonsScreenTest.kt
@@ -177,6 +177,11 @@ class FakeLessonDao(private val flow: MutableStateFlow<List<LessonWithStudent>>)
             flow.map { list -> list.filter { it.student.id == studentId && it.lesson.ownerId == userId } }
         override fun getLessonsWithStudentsInDateRange(startDate: String, endDate: String, userId: Long): Flow<List<LessonWithStudent>> = flowOf(emptyList())
         override fun getLessonsWithStudentsByStudentAndDateRange(studentId: Long, startDate: String, endDate: String, userId: Long): Flow<List<LessonWithStudent>> = flowOf(emptyList())
+
+        override suspend fun insertGroupLessons(lessons: List<Lesson>): List<Long> {
+            lessons.forEach { insert(it) }
+            return lessons.map { it.id }
+        }
     }
 
     class FakeGroupDao : GroupDao {
