+++ app/src/test/java/gr/eduinvoice/ui/revenue/RevenueViewModelTest.kt
@@ -165,6 +165,11 @@ class RevenueViewModelTest {
         override fun getLessonsWithStudentsByStudent(studentId: Long, userId: Long): Flow<List<LessonWithStudent>> = flowOf(emptyList())
         override fun getLessonsWithStudentsInDateRange(startDate: String, endDate: String, userId: Long): Flow<List<LessonWithStudent>> = flowOf(emptyList())
         override fun getLessonsWithStudentsByStudentAndDateRange(studentId: Long, startDate: String, endDate: String, userId: Long): Flow<List<LessonWithStudent>> = flowOf(emptyList())
+
+        override suspend fun insertGroupLessons(lessons: List<Lesson>): List<Long> {
+            lessons.forEach { insert(it) }
+            return lessons.map { it.id }
+        }
     }
 
     class FakeGroupDao : GroupDao {
