+++ data/src/main/java/gr/eduinvoice/data/repository/TutorBillingRepository.kt
@@ -5,14 +5,8 @@ import gr.eduinvoice.data.dao.StudentDao
 import gr.eduinvoice.data.model.Lesson
 import gr.eduinvoice.data.model.Student
 import gr.eduinvoice.data.database.LessonWithStudent
-import gr.eduinvoice.data.model.calculateFee
 import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.combine
 import kotlinx.coroutines.flow.first
-import kotlinx.coroutines.flow.map
-import java.time.DayOfWeek
-import java.time.LocalDate
-import java.time.temporal.TemporalAdjusters
 import javax.inject.Inject
 import javax.inject.Singleton
 
@@ -21,8 +15,7 @@ import javax.inject.Singleton
  *
  * This class sits between the UI layer (ViewModels) and the data layer (DAOs),
  * providing a clean API that hides the complexity of data operations.
- * It offers methods to add, update, delete, and query students and lessons,
- * as well as utility functions for financial calculations.
+ * It offers methods to add, update, delete, and query students and lessons.
  *
  * @Inject tells Hilt to automatically provide the DAOs when creating this repository.
  * @Singleton ensures a single instance is used throughout the app.
@@ -98,14 +91,10 @@ class TutorBillingRepository @Inject constructor(
     suspend fun addGroupLesson(groupId: Long, lesson: Lesson, userId: Long): List<Long> {
         require(lesson.durationMinutes > 0) { "Lesson duration must be positive" }
         val students = groupDao.getStudentsForGroup(groupId, userId).first()
-        return students.map { student ->
-            lessonDao.insert(
-                lesson.copy(
-                    studentId = student.id,
-                    groupId = groupId
-                )
-            )
+        val lessons = students.map { student ->
+            lesson.copy(studentId = student.id, groupId = groupId)
         }
+        return lessonDao.insertGroupLessons(lessons)
     }
 
     /**
@@ -137,100 +126,4 @@ class TutorBillingRepository @Inject constructor(
         return lessonDao.getLessonsWithStudentsByStudent(studentId, userId)
     }
 
-    // ===== Financial Calculations =====
-
-    /**
-     * Holds weekly/monthly totals (plus lesson count) for a given student.
-     */
-    data class StudentFinancialSummary(
-        val student: Student,
-        val weekTotal: Double,
-        val monthTotal: Double,
-        val lessonCount: Int
-    )
-
-    /**
-     * Emits a list of [StudentFinancialSummary], one per active student.
-     */
-    fun getStudentFinancialSummaries(userId: Long): Flow<List<StudentFinancialSummary>> {
-        val today = LocalDate.now()
-        val weekStart = today.with(TemporalAdjusters.previousOrSame(DayOfWeek.MONDAY))
-        val weekEnd = today.with(TemporalAdjusters.nextOrSame(DayOfWeek.SUNDAY))
-        val monthStart = today.withDayOfMonth(1)
-        val monthEnd = today.withDayOfMonth(today.lengthOfMonth())
-
-        return combine(
-            studentDao.getAllActiveStudents(userId),
-            lessonDao.getAllLessons(userId)
-        ) { students, lessons ->
-            students.map { student ->
-                val studentLessons = lessons.filter { it.studentId == student.id }
-                val studentWeekLessons = studentLessons.filter { lesson ->
-                    val date = LocalDate.parse(lesson.date)
-                    !date.isBefore(weekStart) && !date.isAfter(weekEnd)
-                }
-                val studentMonthLessons = studentLessons.filter { lesson ->
-                    val date = LocalDate.parse(lesson.date)
-                    !date.isBefore(monthStart) && !date.isAfter(monthEnd)
-                }
-                val weekTotal = studentWeekLessons.sumOf { it.calculateFee(student) }
-                val monthTotal = studentMonthLessons.sumOf { it.calculateFee(student) }
-                StudentFinancialSummary(
-                    student = student,
-                    weekTotal = weekTotal,
-                    monthTotal = monthTotal,
-                    lessonCount = studentMonthLessons.size
-                )
-            }
-        }
-    }
-
-    /**
-     * Gets total earnings across all students for a given date range.
-     */
-    fun getTotalEarningsForDateRange(startDate: LocalDate, endDate: LocalDate, userId: Long): Flow<Double> {
-        return combine(
-            lessonDao.getLessonsInDateRange(startDate.toString(), endDate.toString(), userId),
-            studentDao.getAllActiveStudents(userId)
-        ) { lessons, students ->
-            val studentMap = students.associateBy { it.id }
-            lessons.sumOf { lesson ->
-                val student = studentMap[lesson.studentId] ?: return@sumOf 0.0
-                lesson.calculateFee(student)
-            }
-        }
-    }
-
-    /**
-     * Detailed report structure for a single student.
-     */
-    data class StudentDetailedReport(
-        val student: Student,
-        val totalLessons: Int,
-        val totalHours: Double,
-        val totalEarnings: Double,
-        val averageLessonDuration: Double,
-        val lastLessonDate: LocalDate?
-    )
-
-    /**
-     * Generates a detailed report for a single student.
-     */
-    suspend fun getStudentDetailedReport(studentId: Long, userId: Long): StudentDetailedReport? {
-        val student = studentDao.getStudentById(studentId, userId).first() ?: return null
-        val lessons: List<Lesson> = lessonDao.getLessonsByStudentId(studentId, userId).first()
-        val totalMinutes = lessons.sumOf { it.durationMinutes }
-        val totalHours = totalMinutes / 60.0
-        val totalEarnings = lessons.sumOf { it.calculateFee(student) }
-        val averageDuration = if (lessons.isNotEmpty()) totalMinutes.toDouble() / lessons.size else 0.0
-        val lastLessonDate = lessons.maxOfOrNull { LocalDate.parse(it.date) }
-        return StudentDetailedReport(
-            student = student,
-            totalLessons = lessons.size,
-            totalHours = totalHours,
-            totalEarnings = totalEarnings,
-            averageLessonDuration = averageDuration,
-            lastLessonDate = lastLessonDate
-        )
-    }
 }
